requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2.16
django-cors-headers==3.13.0
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.4
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.1
pluggy==0.13.1
psycopg2==2.9.7
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.26.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.8.0
urllib3==1.26.16
# pip install pyyaml
djoser djangorestframework-simplejwt==4.7.2